Fix duplicate categories not reflected in taskboss error
